Updated V1 commit

+++ preprocess/build_dataset.py
@@ -1,9 +1,16 @@
-import chess.pgn
+from pathlib import Path
 import csv
-from tqdm import tqdm
+import chess.pgn
 
-PGN_FILE = "../data/raw/elite_games.pgn"
-OUTPUT_FILE = "../data/processed/positions.csv"
+PROJECT_ROOT = Path(__file__).resolve().parent.parent
+
+PGN_FILE = PROJECT_ROOT / "data" / "raw" / "elite_games.pgn"
+OUTPUT_FILE = PROJECT_ROOT / "data" / "processed" / "positions.csv"
+
+OUTPUT_FILE.parent.mkdir(parents=True, exist_ok=True)
+
+if not PGN_FILE.exists():
+    raise FileNotFoundError(f"PGN file not found:\n{PGN_FILE}")
 
 with open(PGN_FILE, "r", encoding="utf-8") as pgn, \
      open(OUTPUT_FILE, "w", newline="", encoding="utf-8") as csvfile:
@@ -25,6 +32,7 @@ with open(PGN_FILE, "r", encoding="utf-8") as pgn, \
     position_id = 0
 
     while True:
+
         game = chess.pgn.read_game(pgn)
 
         if game is None:
@@ -38,7 +46,7 @@ with open(PGN_FILE, "r", encoding="utf-8") as pgn, \
 
         for ply, move in enumerate(game.mainline_moves()):
 
-            if ply < 8:          # Skip opening theory
+            if ply < 8:
                 board.push(move)
                 continue
